Make compile warning-free on cygwin
svn:r936

+++ src/common/util.c
@@ -532,7 +532,7 @@ int check_private_dir(const char *dirname, int create)
     return -1;
   }
   if (st.st_uid != getuid()) {
-    log(LOG_WARN, "%s is not owned by this UID (%d)", dirname, getuid());
+    log(LOG_WARN, "%s is not owned by this UID (%d)", dirname, (int)getuid());
     return -1;
   }
   if (st.st_mode & 0077) {
